requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.42.0
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.10.1
django-taggit==1.3.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.6
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
cloudinary=1.23.0
